requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cloudinary==1.24.0
cryptography==3.4.5
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.6
django-allauth==0.44.0
django-cors-headers==3.7.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==19.7.1
idna==2.10
oauthlib==3.1.0
Pillow==8.1.0
pycparser==2.20
PyJWT==2.0.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
virtualenv==20.4.2
whitenoise==5.2.0
